v0.2 Source Code
This is the first ever release of DiscordL, and therefore it has the most primitive features outside of what it was intended to do. However it does its job as indeed, it is fit for basic purpose and it is stable, if there are any issues please raise this up in "Issues".